Moved tesseract .h files (which we use for windows builds) into their own directory so the reference is consistent when building on linux from a tesseract in a standard location.
