+++ windows/libs/include/tesseract/apitypes.h
