+++ windows/libs/include/tesseract/baseapi.h
