+++ windows/libs/include/tesseract/capi.h
