+++ windows/libs/include/tesseract/platform.h
