+++ windows/libs/include/tesseract/renderer.h
